+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.2] - 2022-12-12
+### Added
+- Banner rotate
+### Changed
+- Found banner must match campaign medium 
+
 ## [2.2.1] - 2022-12-08
 ### Fixed
 - Auth redirect chain
@@ -755,6 +761,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0]
 
 [Unreleased]: https://github.com/adshares/adserver/compare/v2.2.1...develop
+[2.2.2]: https://github.com/adshares/adserver/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/adshares/adserver/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/adshares/adserver/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/adshares/adserver/compare/v2.1.0...v2.1.1